+++ NEWS.md
@@ -3,6 +3,7 @@
 * The WebGL code did not handle the case of lit objects
 with no lights properly.  Now they display the emission
 color only.  (Issue #494).
+* Several deprecated functions have been made defunct.
 
 # rgl 1.3.31
 
@@ -18,8 +19,12 @@ has been dropped at the request of CRAN.
 ## Bug fixes
 
 * `arrow3d(type = "extrusion")` was broken by the changes to triangulation in version 1.3.16.
+<<<<<<< Updated upstream
 * Changes last year to `writePLY()` introduced an error
 in some cases (issue #489).
+=======
+* Many deprecated functions have been made defunct.
+>>>>>>> Stashed changes
 
 # rgl 1.3.24